+++ src/engine/chunk.h
@@ -20,31 +20,31 @@ class Chunk
 
         void update(f32 dt);
         void render(Renderer *renderer);
-        void create_cell(u16 x, u16 y, u8 id);
+        void create_cell(i32 x, i32 y, u8 id);
         void apply_draw();
 
         Chunk* get_neighbour(i32 x, i32 y);
         void set_neighbour(Chunk* neighbour);
 
-        Cell get_cell(u16 x, u16 y);
-        void set_cell(u16 x, u16 y, Cell cell);
-        bool is_empty(u16 x, u16 y);
-        bool in_bounds(u16 x, u16 y) { return x >= 0 && x < size && y >= 0 && y < size; };
+        Cell get_cell(i32 x, i32 y);
+        void set_cell(i32 x, i32 y, Cell cell);
+        bool is_empty(i32 x, i32 y);
+        bool in_bounds(i32 x, i32 y) { return x >= 0 && x < size && y >= 0 && y < size; };
         
         vec2i get_world_position() { return {world_x, world_y}; };
         vec2i get_cell_chunk_position(i32 x, i32 y);
 
-        static const u16 size = 120;
-        static const u16 area = size * size;
+        static const u32 size = 120;
+        static const u32 area = size * size;
     private:
         //Helper functions
-        void set_updated(u16 x, u16 y, u8 value) { update_buffer[y * size + x] = value; };
-        bool is_updated(u16 x, u16 y) { return update_buffer[y * size + x] > 0; };
+        void set_updated(i32 x, i32 y, u8 value) { update_buffer[y * size + x] = value; };
+        bool is_updated(i32 x, i32 y) { return update_buffer[y * size + x] > 0; };
         
         //Cell updates
         void update_bodies(bool has_ticked);
-        void update_sand(u16 x, u16 y);
-        void update_water(u16 x, u16 y);
+        void update_sand(i32 x, i32 y);
+        void update_water(i32 x, i32 y);
         
         //Data
         std::array<Cell, area> cells;
